deleted:    lib/src/core/refresh_widget.dart 	modified:   lib/src/features/changes/application/services/changes_service.dart 	modified:   lib/src/features/changes/data/repositories/changes_repository.dart 	modified:   lib/src/features/changes/presentation/controllers/changes_controller.dart 	modified:   lib/src/features/changes/presentation/widgets/changes_page.dart 	renamed:    lib/src/features/changes/presentation/widgets/add_changes_widget.dart -> lib/src/features/changes/presentation/widgets/crud_changes_widget.dart 	modified:   lib/src/features/news/data/repositories/news_repository.dart 	modified:   lib/src/features/schedule/application/services/schedule_service.dart 	modified:   lib/src/features/schedule/data/repositories/schedule_repository.dart 	modified:   lib/src/features/schedule/presentation/controllers/schedule_controller.dart 	renamed:    lib/src/features/schedule/presentation/widgets/add_schedule_page.dart -> lib/src/features/schedule/presentation/widgets/crud_schedule_page.dart 	modified:   lib/src/features/schedule/presentation/widgets/schedule_page.dart

+++ lib/src/features/changes/application/services/changes_service.dart
@@ -1,4 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:rv1g_info/src/constants/const.dart';
 import 'package:rv1g_info/src/features/changes/data/repositories/changes_repository.dart';
 
 import '../../domain/models/changes.dart';
@@ -10,12 +11,12 @@ class ChangesService {
 
   final ChangesRepository changesRepository;
 
-  Future<void> addChanges(String tag, String imagePath) async{
-    return await changesRepository.addChanges(tag, imagePath);
-  }
-
-  Future<void> deleteChanges(String tag, String imageUrl) async {
-    return await changesRepository.deleteChanges(tag, imageUrl);
+  Future<void> updateChanges(String tag, String imagePath, String imageUrl) async{
+    if(imageUrl != noChanges && imagePath == ""){
+      return await changesRepository.deleteChanges(tag, imageUrl);
+    } else if(imagePath != ""){
+      return await changesRepository.updateChanges(tag, imagePath, imageUrl);
+    }
   }
 
   Future<Map<String, Changes>> getChanges() async{
